backend/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
opencv-python==4.10.0.84
numpy==1.26.4
pymongo==4.10.1
python-dotenv==1.0.1
typing_extensions==4.7.1